refactor(skills): address code review comments and reorganize skill logic
- Move skillLoader and skillManager to dedicated 'packages/core/src/skills' directory
- Refactor 'SkillLoader' into standalone functions to align with project conventions
- Document 'SkillDefinition' interface fields (location, body)
- Remove dead code ('seenLocations' set) from 'loadSkillsFromDir'
- Update re-exports and internal imports for consistent usage

Part of https://github.com/google-gemini/gemini-cli/issues/15693
