+++ packages/core/src/skills/skillManager.test.ts
